compose/kafka_stack/src/requirements.txt
python-logstash
flask
redis
requests
